requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
Deprecated==1.2.13
Flask==2.0.2
idna==3.3
importlib-metadata==4.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
pkg_resources==0.0.0
pycparser==2.21
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.5.0
requests==2.27.1
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.7.0
